Move read by page to Readable protocol
