+++ Jayme/Readable.swift
@@ -28,7 +28,7 @@ public protocol Readable: Repository {
 
 public extension Readable {
     
-    /// Fetches the only entity from this repository
+    /// Fetches the only entity from this repository.
     /// Returns a `Future` containing the only entity in the repository, or the relevant `JaymeError` that could occur.
     /// Watch out for a `.failure` case with `JaymeError.entityNotFound`.
     public func read() -> Future<EntityType, JaymeError> {
@@ -38,7 +38,7 @@ public extension Readable {
             .andThen { EntityParser().entity(from: $0) }
     }
     
-    /// Fetches all the entities from this repository
+    /// Fetches all the entities from this repository.
     /// Returns a `Future` containing an array with all the entities in the repository, or the relevant `JaymeError` that could occur.
     public func readAll() -> Future<[EntityType], JaymeError> {
         let path = self.name
@@ -47,7 +47,26 @@ public extension Readable {
             .andThen { EntityParser().entities(from: $0) }
     }
     
-    /// Fetches only the entity from this repository that matches the given `id`
+    /// Fetches entities from this repository in a paginated manner, based on Grape conventions (https://github.com/davidcelis/api-pagination).
+    /// Returns a `Future` containing a tuple with an array with the entities in the repository that belong to the specified `pageNumber`, plus a `PageInfo` object containing relevant pagination information, or the a `JaymeError` in case of a failure.
+    public func read(pageNumber: Int, pageSize: Int) -> Future<([EntityType], PageInfo), JaymeError> {
+        let path = self.name + "?page=\(pageNumber)&per_page=\(pageSize)"
+        var pageInfo: PageInfo?
+        let future = self.backend.future(path: path, method: .GET, parameters: nil)
+            .andThen {
+                pageInfo = $0.1
+                return DataParser().dictionaries(from: $0.0)
+            }
+            .andThen {
+                EntityParser<EntityType>().entities(from: $0)
+            }
+            .map {
+                return ($0, pageInfo!)
+        }
+        return future
+    }
+    
+    /// Fetches only the entity from this repository that matches the given `id`.
     /// Returns a `Future` containing the entity matching the `id`, or the relevant `JaymeError` that could occur.
     /// Watch out for a `.failure` case with `JaymeError.entityNotFound`.
     public func read(id: EntityType.IdentifierType) -> Future<EntityType, JaymeError> {
